Opensearch helper refactoring
