+++ llm_job_app/chalicelib/aws/opensearch.py
@@ -4,63 +4,1682 @@ import os
 from requests_aws4auth import AWS4Auth
 from opensearchpy import OpenSearch, RequestsHttpConnection, AWSV4SignerAuth
 
-region = "us-east-2"
-service = "es"
-credentials = boto3.Session().get_credentials()
-awsauth = AWS4Auth(
-    os.getenv("AWS_ACCOUNT_ID"),
-    os.getenv("AWS_ACCESS_KEY"),
-    region,
-    service,
-    session_token=credentials.token,
-)
 
-# The OpenSearch domain endpoint
+class OpenSearch_custom:
+    def __init__(
+        self, host, service="es", credentials=boto3.Session().get_credentials()
+    ):
+        self.hosts = host
+        self.awsauth = AWS4Auth(
+            os.getenv("AWS_ACCESS_KEY_ID"),
+            os.getenv("AWS_SECRET_ACCESS_KEY"),
+            "us-east-2",
+            service,
+            session_token=credentials.token,
+        )
+        self.client = self._creat_client(self.hosts)
 
-host = "search-ui-test-j4slt3fz7uu5ahsm6vgu3hwqem.us-east-2.es.amazonaws.com"
+    def _creat_client(self, host):
+        return OpenSearch(
+            hosts=[{"host": host, "port": 443}],
+            http_auth=self.awsauth,
+            use_ssl=True,
+            verify_certs=True,
+            connection_class=RequestsHttpConnection,
+        )
 
-client = OpenSearch(
-    hosts=[{"host": host, "port": 443}],
-    http_auth=awsauth,
-    use_ssl=True,
-    verify_certs=True,
-    connection_class=RequestsHttpConnection,
-)
+    def save_to_index(self, dict_data: dict, index_name: str):
+        """Save the data to OpenSearch index
+
+        Args:
+            dict_data (dict): data to be saved
+            index_name (str): index name
+        """
+        try:
+            # Save the dictionary to OpenSearch
+            response = self.client.index(
+                index=index_name, body=dict_data, pipeline="add-timestamp"
+            )
+            # Print the response
+            # print(response)
+        except Exception as e:
+            print(e)
+            print("=========Error saving to index for the following data:=========\n")
+            print(dict_data)
+            print()
+
+    def create_index(self, index_name: str):
+        # Create an index with non-default settings.
+        # https://docs.aws.amazon.com/opensearch-service/latest/developerguide/sizing-domains.html#bp-sharding
+        # shards should correspond to 10-30GB where search latency is objective
+        # should be 30-50GB for write-heavy jobs like log analytics
+        # by default, each index == 5 primary shards + 1 replica/primary == 10x total shards
+        # replica shard == copy of a primary shard
+        # shards are distributed amongst nodes for resilience
+        # index > shards > multiple nodes (assuming you have more than 1)
+        # lower shard count == faster reads
+        # higher shard count == faster writes
+        index_body = {"settings": {"index": {"number_of_shards": 1}}}
+
+        try:
+            if self.client.indices.exists(index=index_name):
+                print(f"Index {index_name} already exists")
+                return
+            response = self.client.indices.create(
+                index_name,
+                body=index_body,
+                # pipeline="add-timestamp",
+            )
+            print("\nCreating index:")
+            print(response)
+        except Exception as e:
+            print(e)
+
+    def search(self, index_name: str, query: str, size: int = 5):
+        """Search the index for the query
+
+        Args:
+            index_name (str): index name
+            query (dict): query
+
+        Returns:
+            dict: search results
+        """
+        try:
+            query_formated = {
+                "size": size,
+                "query": {"multi_match": {"query": query, "fields": ["job.title"]}},
+            }
+            response = self.client.search(index=index_name, body=query_formated)
+            return response
+        except Exception as e:
+            print(e)
 
 
-def save_to_index(dict_data: dict, index_name: str):
-    """Save the data to OpenSearch index
+if __name__ == "__main__":
+    host = "search-ui-test-j4slt3fz7uu5ahsm6vgu3hwqem.us-east-2.es.amazonaws.com"
+    os_client = OpenSearch_custom(host=host)
+    index_name = "jobs-index"
+    # os_client.create_index(index_name=index_name)
+    # job_parsed = {\n  "job_basics": {\n    "title": "Sr Data Engineer",\n    "company": "Royal Bank of Canada",\n    "industry": "Banking/Financial Services",\n    "location": "Toronto, ON, Canada",\n    "years_of_experience_required": "5+ years",\n    "employment_type": "Full-time",\n    "source": "Jobs At RBC",\n    "application_link": "https://jobs.rbc.com/ca/en/job/R-0000076286/Sr-Data-Engineer?utm_campaign=google_jobs_apply&utm_source=google_jobs_apply&utm_medium=organic"\n  },\n  "required_skills": {\n    "programming_languages": ["Python", "SQL"],\n    "data_visualization_tools": [],\n    "databases": ["NoSQL", "SQL databases"],\n    "web_development": {\n      "front_end_technologies": [],\n      "back_end_technologies": []\n    },\n    "devops_tools": [],\n    "statistical_analysis_methods": [],\n    "machine_learning_frameworks": [],\n    "machine_learning_operations": [],\n    "cloud_platforms": ["Azure", "Snowflake"],\n    "big_data_technologies": ["Hadoop", "PySpark"],\n    "natural_language_processing": [],\n    "deep_learning_concepts": [],\n    "data_engineering_tools": ["ETL/ELT workflows"],\n    "large_models": []\n  },\n  "company_culture": [\n    "diversity and inclusion",\n    "equitable workplace",\n    "coaching and mentoring"\n  ],\n  "job_benefits": {\n    "salary_range": "",\n    "salary_currency": "CAD",\n    "paid_time_off": "",\n    "retirement_plan_matching": false,\n    "professional_development_budget": "",\n    "employee_discounts": "",\n    "remote_work_policy": {\n      "fully_remote_option": false,\n      "partial_remote_option": "",\n      "work_from_anywhere_program": ""\n    },\n    "relocation_support": ""\n  }\n}\n```'
+    job_parsed = {
+        "job": {
+            "title": "Senior Machine Learning Engineer",
+            "company": "Everbridge",
+            "industry": "Critical Event Management (CEM) AI/ML",
+            "location": "Anywhere",
+            "years_of_experience_required": "5+ years",
+            "employment_type": "Full-time",
+            "source": "LinkedIn",
+            "application_link": "https://ca.linkedin.com/jobs/view/senior-machine-learning-engineer-at-everbridge-3768369146?utm_campaign=google_jobs_apply&utm_source=google_jobs_apply&utm_medium=organic",
+        },
+        "skills": {
+            "programming_languages": ["Python", "Java"],
+            "machine_learning_frameworks": ["PyTorch", "TensorFlow"],
+            "data_engineering_tools": ["Airflow"],
+            "devops_tools": ["Docker", "Kubernetes"],
+            "big_data_technologies": ["Spark"],
+            "natural_language_processing": ["BERT", "BART", "DistilBERT"],
+            "distributed_data_processing_frameworks": ["Spark", "Storm"],
+            "gpu_computing": ["CUDA", "cuDNN"],
+        },
+        "company_culture": [
+            "resilient organizations",
+            "intelligent automation",
+            "comprehensive risk data",
+        ],
+        "job_benefits": {
+            "salary_range": "Not specified",
+            "salary_currency": "Not specified",
+            "paid_time_off": "Not specified",
+            "retirement_plan_matching": False,
+            "professional_development_budget": "Not specified",
+            "employee_discounts": "Not specified",
+            "remote_work_policy": {
+                "fully_remote_option": True,
+                "partial_remote_option": "Not specified",
+                "work_from_anywhere_program": "Not specified",
+            },
+            "relocation_support": False,
+        },
+        "embedding": [
+            -0.010497848503291607,
+            -0.018889375030994415,
+            0.004148506559431553,
+            -0.03704963997006416,
+            2.9746715881628916e-05,
+            0.007270856760442257,
+            -0.02083367109298706,
+            -0.006281831301748753,
+            -0.03731968253850937,
+            -0.02634251117706299,
+            0.017566174268722534,
+            0.01049109734594822,
+            -0.006464109290391207,
+            0.005883520934730768,
+            -0.001890287734568119,
+            0.001323201460763812,
+            0.011348477564752102,
+            -0.004310531076043844,
+            0.0076421634294092655,
+            -0.010781391523778439,
+            -0.02391214109957218,
+            -0.004425298888236284,
+            0.01676955260336399,
+            -0.018052248284220695,
+            -0.011503751389682293,
+            0.010072533041238785,
+            0.018808363005518913,
+            -0.03594247251749039,
+            -0.02832731232047081,
+            5.007232175557874e-06,
+            0.03251295164227486,
+            -0.0025350109208375216,
+            -0.014366187155246735,
+            -0.010099537670612335,
+            -0.014339183457195759,
+            -0.017282631248235703,
+            0.0056134797632694244,
+            -0.002189020859077573,
+            0.013508806936442852,
+            -0.007034570910036564,
+            0.03251295164227486,
+            0.029704522341489792,
+            -0.0062717050313949585,
+            -0.023844629526138306,
+            -0.0025721415877342224,
+            0.02711212821304798,
+            -0.010160297155380249,
+            -0.013191508129239082,
+            -0.016526516526937485,
+            0.03750871121883392,
+            0.002288598334416747,
+            0.01795773394405842,
+            -0.005458205938339233,
+            -0.008587307296693325,
+            -0.007817690260708332,
+            0.024938296526670456,
+            -0.022183876484632492,
+            0.032323919236660004,
+            0.005701242946088314,
+            -0.015608375892043114,
+            0.01266492810100317,
+            0.007365371100604534,
+            -0.008621062152087688,
+            -0.015257323160767555,
+            -0.017255626618862152,
+            0.000571727694477886,
+            -0.006754403468221426,
+            0.03178383782505989,
+            0.004384792409837246,
+            -0.015054792165756226,
+            0.028030266985297203,
+            0.0011797421611845493,
+            0.01337378565222025,
+            0.023871634155511856,
+            0.015986433252692223,
+            -0.011834551580250263,
+            0.0029366970993578434,
+            0.009930761530995369,
+            0.006484362296760082,
+            0.01668854057788849,
+            -0.0002575938997324556,
+            -0.02634251117706299,
+            0.0012843830045312643,
+            0.001643031369894743,
+            0.004506310913711786,
+            0.02291298843920231,
+            -0.00797971524298191,
+            -0.004901246167719364,
+            0.006558623630553484,
+            -0.009782238863408566,
+            -0.0008181401644833386,
+            0.003513909876346588,
+            0.026329008862376213,
+            0.006869170814752579,
+            -0.015716392546892166,
+            0.010464092716574669,
+            -0.007021069061011076,
+            0.0030430257320404053,
+            -0.01564888283610344,
+            -0.032377928495407104,
+            -0.01548685785382986,
+            0.011638771742582321,
+            -0.0203475970774889,
+            -0.003216864774003625,
+            -0.02962351031601429,
+            0.005728247109800577,
+            -0.0028708744794130325,
+            0.006973811890929937,
+            0.0192809347063303,
+            0.005437952931970358,
+            -0.018160264939069748,
+            0.026153482496738434,
+            0.008344270288944244,
+            -0.049174487590789795,
+            -0.024344205856323242,
+            -0.030298613011837006,
+            0.006335839629173279,
+            -0.026950102299451828,
+            -0.0026767824310809374,
+            0.0047864788211882114,
+            0.03769773989915848,
+            0.02789524756371975,
+            0.016175461933016777,
+            -0.03251295164227486,
+            0.03640154004096985,
+            0.02022607997059822,
+            -0.01633748784661293,
+            -0.028894398361444473,
+            -0.0071223340928554535,
+            -0.0003875511756632477,
+            0.013907116837799549,
+            -0.010815146379172802,
+            0.00471221748739481,
+            -0.004135004710406065,
+            -0.009606712497770786,
+            -0.0034362731967121363,
+            -0.027274152263998985,
+            0.003046401310712099,
+            -0.022143371403217316,
+            0.006197443697601557,
+            0.005552720278501511,
+            0.010551856830716133,
+            0.0010675062658265233,
+            -0.005100401584059,
+            0.00664301123470068,
+            -0.0005097026005387306,
+            0.015311331488192081,
+            0.01884886994957924,
+            0.032782990485429764,
+            0.002654841635376215,
+            -0.020725654438138008,
+            -0.028462333604693413,
+            -0.014055639505386353,
+            0.013826104812324047,
+            0.0030869075562804937,
+            -0.014123150147497654,
+            0.019807515665888786,
+            0.01861933432519436,
+            -0.010686877183616161,
+            0.016432002186775208,
+            -0.006393223535269499,
+            -0.004337535239756107,
+            0.007493640761822462,
+            -0.013279271312057972,
+            -0.006558623630553484,
+            0.015027787536382675,
+            0.03513234853744507,
+            -0.007149338256567717,
+            0.0052489242516458035,
+            0.013650578446686268,
+            -0.012691931799054146,
+            0.013670830987393856,
+            -0.04388168081641197,
+            0.01737714558839798,
+            0.03904794529080391,
+            0.0052185445092618465,
+            0.01867334358394146,
+            -0.016148459166288376,
+            -0.035429392009973526,
+            -0.00880334060639143,
+            0.014271672815084457,
+            -0.013414292596280575,
+            0.0160809475928545,
+            0.025869937613606453,
+            -0.016404997557401657,
+            0.001758642727509141,
+            -0.007763681933283806,
+            -0.0016810059314593673,
+            0.005390695761889219,
+            -0.020091058686375618,
+            -0.0004118126817047596,
+            0.02789524756371975,
+            -0.004989009350538254,
+            0.0010320633882656693,
+            -0.6433459520339966,
+            -0.038831911981105804,
+            0.008418532088398933,
+            -0.025775423273444176,
+            0.007574653252959251,
+            0.015257323160767555,
+            -0.00856030359864235,
+            0.009161144495010376,
+            -0.0004213063220959157,
+            0.03948001191020012,
+            -0.005238797515630722,
+            0.027611704543232918,
+            -7.199338142527267e-05,
+            0.012759442441165447,
+            0.008897854946553707,
+            -0.012651425786316395,
+            0.010666623711585999,
+            -0.023588091135025024,
+            -1.842028541432228e-05,
+            0.009856500662863255,
+            -0.006278455723077059,
+            0.01752566732466221,
+            0.004475931636989117,
+            -0.010430337861180305,
+            -0.004799980670213699,
+            0.014366187155246735,
+            0.011213457211852074,
+            -0.009903757832944393,
+            0.006737525574862957,
+            0.015095298178493977,
+            -0.030649665743112564,
+            0.019699499011039734,
+            0.013826104812324047,
+            -0.014933273196220398,
+            0.04083021730184555,
+            0.015014286153018475,
+            -0.009410932660102844,
+            0.020158568397164345,
+            0.025802427902817726,
+            0.031999871134757996,
+            -0.023061510175466537,
+            -0.0011586451437324286,
+            -0.019699499011039734,
+            0.024884289130568504,
+            -0.00035801544436253607,
+            0.024222686886787415,
+            0.009316418319940567,
+            0.010612615384161472,
+            -0.014015133492648602,
+            0.005005887243896723,
+            0.007729926612228155,
+            0.008789838291704655,
+            -0.008114735595881939,
+            0.003915596287697554,
+            0.014852261170744896,
+            -0.007115583401173353,
+            0.01755267195403576,
+            -0.02553238719701767,
+            -0.0042362697422504425,
+            -0.01913241297006607,
+            0.012556911446154118,
+            0.01011303998529911,
+            -0.0256809089332819,
+            0.006558623630553484,
+            -0.016607528552412987,
+            0.003908845130354166,
+            0.008729078806936741,
+            -0.016810059547424316,
+            -0.009370426647365093,
+            -0.031513798981904984,
+            0.010538354516029358,
+            0.010356076993048191,
+            0.032161895185709,
+            -0.024155177175998688,
+            0.008756083436310291,
+            0.030325617641210556,
+            0.047068167477846146,
+            0.011551008559763432,
+            -0.01348855346441269,
+            0.002724039601162076,
+            0.01948346570134163,
+            0.005117279011756182,
+            -0.0038919674698263407,
+            -0.014069141820073128,
+            0.02311551943421364,
+            -0.010437089018523693,
+            -0.016634533181786537,
+            -0.021765314042568207,
+            0.005961157381534576,
+            0.0030767808202654123,
+            0.0278682429343462,
+            0.02622099220752716,
+            -0.013090242631733418,
+            -0.049984611570835114,
+            0.0023037882056087255,
+            0.010875905863940716,
+            0.006322337314486504,
+            -0.005684365518391132,
+            0.008641315624117851,
+            -0.023993153125047684,
+            -0.010713880881667137,
+            -0.009491944685578346,
+            0.029812538996338844,
+            -0.006926554720848799,
+            0.01890287734568119,
+            0.0076016574166715145,
+            -0.011733286082744598,
+            0.01221936009824276,
+            0.009127389639616013,
+            0.0009915572591125965,
+            -0.005599977448582649,
+            -0.014217664487659931,
+            -0.015257323160767555,
+            0.00848604179918766,
+            0.0025113823357969522,
+            -0.027544192969799042,
+            -0.01184130273759365,
+            -0.007048073224723339,
+            0.002649778500199318,
+            -0.018227774649858475,
+            -0.020091058686375618,
+            0.012388136237859726,
+            0.014960277825593948,
+            -0.008837095461785793,
+            0.005846390035003424,
+            0.024317201226949692,
+            0.011632020585238934,
+            0.005545969121158123,
+            0.00497888308018446,
+            0.01936194859445095,
+            0.011476747691631317,
+            0.01861933432519436,
+            0.01230712328106165,
+            -0.0010514726163819432,
+            -0.005083524156361818,
+            0.0016261538257822394,
+            0.019145915284752846,
+            -0.004624453838914633,
+            -0.004162008408457041,
+            -0.03829183056950569,
+            -0.01135522872209549,
+            -0.0171206071972847,
+            0.00139408721588552,
+            0.002263281960040331,
+            -0.01956447772681713,
+            -0.0063122110441327095,
+            -0.03680660203099251,
+            0.02273746207356453,
+            0.011740037240087986,
+            -0.00545483035966754,
+            -0.005350189749151468,
+            0.0033721383661031723,
+            -0.01668854057788849,
+            0.03170282766222954,
+            0.0043915435671806335,
+            -0.011037930846214294,
+            0.003277624025940895,
+            -0.026842085644602776,
+            -0.018119757995009422,
+            -0.02083367109298706,
+            -0.012199107557535172,
+            0.005400822032243013,
+            -0.038399845361709595,
+            -0.005937528796494007,
+            -0.005606728605926037,
+            -0.006960309576243162,
+            0.008188996464014053,
+            0.005083524156361818,
+            -0.004651458002626896,
+            -0.024681758135557175,
+            0.008789838291704655,
+            -0.01809275522828102,
+            0.010315570048987865,
+            0.02622099220752716,
+            -0.03148679435253143,
+            0.018160264939069748,
+            -0.02708512358367443,
+            0.010167048312723637,
+            0.00643372954800725,
+            -0.017282631248235703,
+            -0.0014725679066032171,
+            0.012624422088265419,
+            -0.015837911516427994,
+            0.0008556927787140012,
+            0.04412471875548363,
+            0.0048944950103759766,
+            -0.00557634886354208,
+            0.01737714558839798,
+            -0.006933305412530899,
+            -0.012158600613474846,
+            -0.010470843873918056,
+            0.00797971524298191,
+            -0.011037930846214294,
+            0.006011790130287409,
+            -0.012523156590759754,
+            0.01328602246940136,
+            -0.0070818280801177025,
+            -0.00815524160861969,
+            -0.008965364657342434,
+            -0.00543120177462697,
+            0.004766225814819336,
+            0.0247492678463459,
+            0.027409173548221588,
+            0.0013358596479520202,
+            0.014298676513135433,
+            -0.024600746110081673,
+            0.005559471435844898,
+            -0.026207489892840385,
+            0.009701226837933064,
+            0.02827330492436886,
+            0.01064637117087841,
+            -0.008756083436310291,
+            -0.007385624572634697,
+            -0.0012042145244777203,
+            -0.015594874508678913,
+            0.0046008252538740635,
+            -0.012921467423439026,
+            0.013826104812324047,
+            -0.0278682429343462,
+            -0.0049248747527599335,
+            0.011908813379704952,
+            -0.013387287966907024,
+            0.01567588560283184,
+            -0.005853141192346811,
+            0.003554416121914983,
+            0.02273746207356453,
+            -0.011618519201874733,
+            0.02063114009797573,
+            -0.01246239710599184,
+            -0.031513798981904984,
+            0.008506295271217823,
+            0.01093666534870863,
+            0.029731526970863342,
+            0.011665776371955872,
+            0.0246682558208704,
+            0.0192809347063303,
+            0.018781360238790512,
+            -0.038345836102962494,
+            0.03591546788811684,
+            -0.00999827217310667,
+            -0.00516453618183732,
+            0.016067445278167725,
+            0.02400665543973446,
+            -0.013880113139748573,
+            0.02149527333676815,
+            -0.0050801485776901245,
+            0.03812980651855469,
+            0.016756050288677216,
+            -0.01344129629433155,
+            0.003939224872738123,
+            -0.019875025376677513,
+            -0.009647218510508537,
+            0.0009510510135442019,
+            0.006575501058250666,
+            0.017674190923571587,
+            -0.002594082383438945,
+            0.016607528552412987,
+            0.004435425158590078,
+            0.01907840371131897,
+            0.017701195552945137,
+            0.0024337454233318567,
+            0.014339183457195759,
+            0.026112975552678108,
+            0.003275936236605048,
+            0.016013437882065773,
+            -0.0006261578528210521,
+            -0.011375482194125652,
+            -0.008216001093387604,
+            -8.359671483049169e-05,
+            -0.01224636472761631,
+            -0.006197443697601557,
+            -0.026153482496738434,
+            0.008526547811925411,
+            -0.024128172546625137,
+            -0.0024050537031143904,
+            -0.0061670639552176,
+            0.0024303700774908066,
+            0.017822712659835815,
+            -0.008668319322168827,
+            0.01956447772681713,
+            -0.002594082383438945,
+            -0.01676955260336399,
+            0.02075265906751156,
+            0.006008414551615715,
+            -0.008614311926066875,
+            0.00795271061360836,
+            -0.015837911516427994,
+            0.008958613499999046,
+            -0.026477530598640442,
+            -0.009937512688338757,
+            0.008337519131600857,
+            0.014960277825593948,
+            -0.007372122257947922,
+            0.025883439928293228,
+            -0.01151050254702568,
+            0.003810955211520195,
+            0.025478379800915718,
+            -0.006460733711719513,
+            -0.0039999838918447495,
+            -0.01159826572984457,
+            0.010410084389150143,
+            0.00897886697202921,
+            -0.019726503640413284,
+            -0.006126557942479849,
+            0.055628467351198196,
+            -0.0019949288107454777,
+            -0.017066597938537598,
+            -0.01032907236367464,
+            0.02175181172788143,
+            -0.02792225033044815,
+            -0.006092802621424198,
+            -0.00013512604346033186,
+            0.019267432391643524,
+            0.005616855341941118,
+            0.013961125165224075,
+            -0.01119320373982191,
+            0.00893160980194807,
+            0.01171303354203701,
+            0.015540866181254387,
+            0.015999935567378998,
+            -0.010747636668384075,
+            -0.015864916145801544,
+            -0.006038794293999672,
+            0.037157658487558365,
+            0.05020064115524292,
+            0.009215152822434902,
+            -0.009100385010242462,
+            0.004546817392110825,
+            -0.021171223372220993,
+            -0.01351555809378624,
+            -0.03750871121883392,
+            -0.0027375416830182076,
+            0.01544635184109211,
+            -0.019226927310228348,
+            -0.01117970235645771,
+            0.009073381312191486,
+            -0.0005924027063883841,
+            -0.014501207508146763,
+            0.01636449061334133,
+            -0.002258218824863434,
+            0.01070712972432375,
+            -0.018835367634892464,
+            -0.00385483680292964,
+            -0.0013147626304998994,
+            0.0012978850863873959,
+            -0.002324041211977601,
+            0.005940904375165701,
+            -0.013076741248369217,
+            0.03294501453638077,
+            0.008897854946553707,
+            0.03135177120566368,
+            0.020496120676398277,
+            -0.001726575312204659,
+            -0.04153232276439667,
+            -0.019010894000530243,
+            0.032891009002923965,
+            0.01786321960389614,
+            0.0006329088355414569,
+            0.000529955665115267,
+            0.06146135553717613,
+            0.024627748876810074,
+            0.028894398361444473,
+            0.01740415021777153,
+            -0.025316353887319565,
+            0.0150007838383317,
+            0.013049736618995667,
+            0.004901246167719364,
+            -0.009431186132133007,
+            0.018605832010507584,
+            -0.02377711981534958,
+            -0.00708857923746109,
+            0.03348509967327118,
+            -0.004820233676582575,
+            -0.019240429624915123,
+            0.028678366914391518,
+            0.012921467423439026,
+            -0.048796430230140686,
+            -0.013292773626744747,
+            0.02685558795928955,
+            0.01459572184830904,
+            0.01895688660442829,
+            -0.004087747540324926,
+            -0.023034507408738136,
+            -0.04471880942583084,
+            -0.005097026005387306,
+            -0.022845478728413582,
+            -0.00984299834817648,
+            -0.006818538065999746,
+            -0.009802492335438728,
+            0.010599114000797272,
+            0.006649762392044067,
+            0.009586459025740623,
+            -0.021319745108485222,
+            0.009978018701076508,
+            -0.02017207071185112,
+            -0.02604546584188938,
+            -0.028003262355923653,
+            0.023588091135025024,
+            0.026126477867364883,
+            -0.008864099159836769,
+            0.015324832871556282,
+            0.0047155930660665035,
+            0.0003989435499534011,
+            0.0010801644530147314,
+            -0.027139130979776382,
+            -0.0076016574166715145,
+            -0.004577196668833494,
+            -0.008310515433549881,
+            -0.023304548114538193,
+            -0.02291298843920231,
+            -0.02492479421198368,
+            0.006524868309497833,
+            -0.009188149124383926,
+            0.02164379507303238,
+            0.0075341472402215,
+            -0.014096146449446678,
+            -0.011996576562523842,
+            0.00909363478422165,
+            0.0086008096113801,
+            0.0139341214671731,
+            0.020941687747836113,
+            0.0038919674698263407,
+            -0.0007367059006355703,
+            -0.018349293619394302,
+            -0.007716424763202667,
+            -0.03683360666036606,
+            -0.029029419645667076,
+            -0.04690613970160484,
+            -0.013961125165224075,
+            0.015540866181254387,
+            0.018983889371156693,
+            0.017485162243247032,
+            0.00027404952561482787,
+            0.014285175129771233,
+            0.029272455722093582,
+            0.002619398757815361,
+            0.026383016258478165,
+            0.010362827219069004,
+            0.0014278424205258489,
+            -0.009626965969800949,
+            0.0063695949502289295,
+            -0.0066193826496601105,
+            0.02555939182639122,
+            -0.008047224953770638,
+            0.021373754367232323,
+            -0.03559141978621483,
+            0.026085970923304558,
+            0.04034414142370224,
+            0.011382232420146465,
+            0.02599145658314228,
+            0.02057713270187378,
+            -0.008621062152087688,
+            -0.040128108114004135,
+            -0.008020221255719662,
+            -0.015189812518656254,
+            0.03696862980723381,
+            -0.015297829173505306,
+            -0.011780543252825737,
+            -0.007311363238841295,
+            -0.016553519293665886,
+            -0.017066597938537598,
+            0.006805036216974258,
+            -0.023790622130036354,
+            -0.021333247423171997,
+            -0.01564888283610344,
+            -0.0012244676472619176,
+            -0.0057113696821033955,
+            -0.0022447167430073023,
+            0.008668319322168827,
+            -0.031054727733135223,
+            -0.010761138051748276,
+            -0.0007430349942296743,
+            -0.012118094600737095,
+            0.03426821529865265,
+            0.0026987234596163034,
+            0.025370363146066666,
+            -0.008101233281195164,
+            -0.005556095857173204,
+            -0.0112674655392766,
+            -0.0015510485973209143,
+            0.004202514886856079,
+            -0.012887711636722088,
+            0.03002857230603695,
+            0.04409771412611008,
+            0.024195684120059013,
+            -0.01910540834069252,
+            0.034538257867097855,
+            0.01740415021777153,
+            -0.001493664924055338,
+            0.005876769777387381,
+            0.011787294410169125,
+            0.008236253634095192,
+            -0.017728198319673538,
+            -0.01044384017586708,
+            0.02403365820646286,
+            0.0002871296601369977,
+            -0.003348509781062603,
+            -0.030595658347010612,
+            -0.0009983081836253405,
+            0.00012658177001867443,
+            -0.010713880881667137,
+            0.0035274119582027197,
+            -0.028948407620191574,
+            -0.02098219469189644,
+            -0.011213457211852074,
+            0.028030266985297203,
+            -0.00590039836242795,
+            0.013130748644471169,
+            -0.010700379498302937,
+            -0.00038923893589526415,
+            0.033242061734199524,
+            0.012975475750863552,
+            0.00287762563675642,
+            0.01654001884162426,
+            0.0036354283802211285,
+            0.000544301641639322,
+            0.007250603754073381,
+            0.014096146449446678,
+            0.02406066283583641,
+            -0.01483875885605812,
+            -0.014447199180722237,
+            -0.041019245982170105,
+            -0.028246300294995308,
+            -0.011814299039542675,
+            0.009815994650125504,
+            0.026531539857387543,
+            0.006335839629173279,
+            -0.003956102300435305,
+            -0.017012590542435646,
+            0.02884039096534252,
+            -0.02271045744419098,
+            0.010956917889416218,
+            0.0007797437137924135,
+            -0.03656356781721115,
+            0.006453982554376125,
+            -0.03175683319568634,
+            -0.005839638877660036,
+            -0.016391495242714882,
+            -0.006865795236080885,
+            0.009235406294465065,
+            0.009734981693327427,
+            0.033269066363573074,
+            -0.022548433393239975,
+            -0.020671647042036057,
+            0.0028708744794130325,
+            0.005032891407608986,
+            0.05414324253797531,
+            0.004614327568560839,
+            0.02495179884135723,
+            0.03893992677330971,
+            0.020064054057002068,
+            0.00022426070063374937,
+            -0.002258218824863434,
+            0.03140578046441078,
+            -0.011456494219601154,
+            0.02728765457868576,
+            0.04034414142370224,
+            0.009924011304974556,
+            -0.011679277755320072,
+            -0.015392343513667583,
+            0.012583916075527668,
+            0.00041392238927073777,
+            -0.021454766392707825,
+            0.032891009002923965,
+            -0.02665305696427822,
+            -0.012205857783555984,
+            -0.01124046090990305,
+            -0.013265769928693771,
+            -0.03386315703392029,
+            0.01548685785382986,
+            -0.022521428763866425,
+            0.005390695761889219,
+            -0.007466636598110199,
+            -0.006778032053261995,
+            0.008148490451276302,
+            0.0012194043956696987,
+            -0.01044384017586708,
+            0.002815178595483303,
+            0.029110431671142578,
+            -0.0041383798234164715,
+            0.01397462747991085,
+            -0.021454766392707825,
+            -0.012448894791305065,
+            0.013839607127010822,
+            0.0021704554092139006,
+            0.02921844832599163,
+            -0.009451438672840595,
+            0.006048921030014753,
+            0.01095016673207283,
+            0.024330703541636467,
+            0.005070021841675043,
+            -0.0010810083476826549,
+            0.023304548114538193,
+            0.03723866865038872,
+            -0.017161112278699875,
+            -0.021792316809296608,
+            0.008087730966508389,
+            -0.028219295665621758,
+            0.011132445186376572,
+            -0.004310531076043844,
+            -0.016850564628839493,
+            -0.01541934721171856,
+            0.001859908108599484,
+            -0.003571293782442808,
+            0.019199922680854797,
+            0.01017379853874445,
+            -0.01999654434621334,
+            -0.01053160335868597,
+            -0.012691931799054146,
+            -0.004259898327291012,
+            0.03145978972315788,
+            -0.007230350747704506,
+            0.022642947733402252,
+            -0.01676955260336399,
+            -0.0042531476356089115,
+            0.012367882765829563,
+            -0.004627829417586327,
+            0.007871698588132858,
+            0.0048843687400221825,
+            0.005205042194575071,
+            -0.0031291013583540916,
+            0.01806575059890747,
+            -0.020874178037047386,
+            0.018079252913594246,
+            -0.01676955260336399,
+            0.003932473715394735,
+            -0.017674190923571587,
+            0.016121454536914825,
+            -0.030298613011837006,
+            -0.00370631436817348,
+            0.015608375892043114,
+            -0.019253931939601898,
+            -0.013684333302080631,
+            0.012226111255586147,
+            -0.002843870548531413,
+            0.0017383896047249436,
+            0.001504635321907699,
+            -0.008985618129372597,
+            0.03599648177623749,
+            0.0018514693947508931,
+            0.00477635208517313,
+            -0.013117247261106968,
+            -0.003837959375232458,
+            0.019145915284752846,
+            -0.03556441515684128,
+            -0.002546825213357806,
+            -0.005606728605926037,
+            -0.022467421367764473,
+            -0.018376298248767853,
+            -0.015378841198980808,
+            0.0007341742748394608,
+            -0.0182682815939188,
+            -0.03821081668138504,
+            -0.03359311446547508,
+            0.013765345327556133,
+            0.005856516771018505,
+            0.00013976737682241946,
+            -0.0007767901406623423,
+            -0.018916379660367966,
+            0.004455678164958954,
+            -0.02215687371790409,
+            0.028138283640146255,
+            -0.017255626618862152,
+            -0.000892823445610702,
+            0.011807547882199287,
+            0.03615850582718849,
+            0.040965236723423004,
+            0.022872481495141983,
+            -0.023588091135025024,
+            -0.014447199180722237,
+            -0.021913835778832436,
+            -0.02258893847465515,
+            -0.0342952199280262,
+            0.008938360959291458,
+            -0.03183784708380699,
+            0.045285895466804504,
+            0.00988350436091423,
+            -0.005407573189586401,
+            0.006393223535269499,
+            -0.03426821529865265,
+            -0.007075076922774315,
+            -0.02083367109298706,
+            0.011524004861712456,
+            0.030784687027335167,
+            0.017687693238258362,
+            0.004114751238375902,
+            0.0006721491809003055,
+            -0.0013291086070239544,
+            -0.0018042122246697545,
+            0.027476683259010315,
+            0.0007594906492158771,
+            -0.004273400641977787,
+            0.009984769858419895,
+            -0.0035307875368744135,
+            -0.001491977134719491,
+            0.007169591728597879,
+            -0.036239515990018845,
+            -0.019172918051481247,
+            0.01806575059890747,
+            -0.004749347921460867,
+            -0.008053976111114025,
+            0.0007037946488708258,
+            -0.01090966071933508,
+            -0.01479825284332037,
+            0.0008666631765663624,
+            -0.00649111345410347,
+            -0.007999967783689499,
+            -0.00370968971401453,
+            0.005626981612294912,
+            -0.0019358572317287326,
+            -0.00835777260363102,
+            -0.0038615879602730274,
+            -0.022426914423704147,
+            0.018916379660367966,
+            -0.010659872554242611,
+            0.006484362296760082,
+            -0.0018514693947508931,
+            0.021738309413194656,
+            -0.00833076797425747,
+            0.0015797405503690243,
+            0.007270856760442257,
+            0.0013080115895718336,
+            0.028057271614670753,
+            0.0032894383184611797,
+            0.0061569372192025185,
+            0.01755267195403576,
+            -0.002057375619187951,
+            0.0018464061431586742,
+            -0.004172135144472122,
+            -0.02677457593381405,
+            -0.027382168918848038,
+            -0.015432849526405334,
+            0.0016227783635258675,
+            -0.023939143866300583,
+            -0.01948346570134163,
+            -0.012050584889948368,
+            0.028570350259542465,
+            0.001228687004186213,
+            -0.012226111255586147,
+            -0.009282663464546204,
+            -0.02083367109298706,
+            -0.007432881742715836,
+            0.014960277825593948,
+            0.008324016816914082,
+            -0.012624422088265419,
+            -0.02011806331574917,
+            -0.007061575073748827,
+            -0.014325681142508984,
+            -0.015473355539143085,
+            -0.022062359377741814,
+            0.002523196628317237,
+            -0.02268345281481743,
+            0.005637108348309994,
+            0.015365338884294033,
+            -0.012367882765829563,
+            -0.025748420506715775,
+            0.0073181139305233955,
+            0.021778816357254982,
+            -0.008384776301681995,
+            0.008080979809165001,
+            0.19054101407527924,
+            -0.021427761763334274,
+            0.014892767183482647,
+            -0.0008236253634095192,
+            0.005556095857173204,
+            0.02019907534122467,
+            0.022750964388251305,
+            0.008324016816914082,
+            -0.010335823521018028,
+            0.0014717240119352937,
+            -0.008290261961519718,
+            0.005059895105659962,
+            -0.015810906887054443,
+            -0.0063527170568704605,
+            -9.09806476556696e-05,
+            -0.02212986908853054,
+            -0.03645554929971695,
+            -0.025289349257946014,
+            -0.018578829243779182,
+            -0.0007042165962047875,
+            0.0031020971946418285,
+            0.016269976273179054,
+            -0.02881338633596897,
+            -0.028435328975319862,
+            -0.010450591333210468,
+            0.01633748784661293,
+            0.020712153986096382,
+            0.016094449907541275,
+            0.022237885743379593,
+            0.01948346570134163,
+            -0.002025308320298791,
+            0.007999967783689499,
+            -0.011375482194125652,
+            -0.012091090902686119,
+            -0.00557634886354208,
+            -0.021400758996605873,
+            0.011632020585238934,
+            -0.0182142723351717,
+            0.02316952683031559,
+            -0.006602505221962929,
+            0.004003359470516443,
+            -0.017242124304175377,
+            0.02919144369661808,
+            -0.025316353887319565,
+            0.003621926298364997,
+            0.01783621497452259,
+            -0.03507833927869797,
+            -0.000144936129800044,
+            -0.0018666591495275497,
+            0.021508773788809776,
+            -0.021994847804307938,
+            0.002686909167096019,
+            0.04134329408407211,
+            0.025572894141077995,
+            0.006595754064619541,
+            0.017188116908073425,
+            0.012097842060029507,
+            0.022237885743379593,
+            0.004378041718155146,
+            0.019523972645401955,
+            -0.013751843944191933,
+            0.016904573887586594,
+            -0.0035037833731621504,
+            0.023237038403749466,
+            0.021387256681919098,
+            -0.0031729829497635365,
+            -0.019240429624915123,
+            0.013542561791837215,
+            -0.001013498054817319,
+            -0.0017873345641419291,
+            -0.02997456304728985,
+            -0.013920619152486324,
+            0.011571262031793594,
+            0.00231560249812901,
+            -0.02657204493880272,
+            -0.029758529737591743,
+            0.019321441650390625,
+            0.014163656160235405,
+            0.004070869646966457,
+            0.05068671703338623,
+            -0.008303764276206493,
+            -0.031594809144735336,
+            0.0016202466795220971,
+            -0.020442111417651176,
+            0.004084371961653233,
+            -0.006710521411150694,
+            0.018362795934081078,
+            -0.017269128933548927,
+            -0.007527396082878113,
+            -0.0107341343536973,
+            -0.008837095461785793,
+            -0.025896942242980003,
+            -0.013009230606257915,
+            0.00038712925743311644,
+            0.012300373055040836,
+            0.02743617631494999,
+            0.007540897931903601,
+            0.017269128933548927,
+            -0.0023426066618412733,
+            -0.008182245306670666,
+            -0.03318805247545242,
+            0.05778879672288895,
+            0.03812980651855469,
+            0.013259018771350384,
+            -0.016877569258213043,
+            0.00017035796190612018,
+            0.008290261961519718,
+            0.0030075828544795513,
+            0.009289414621889591,
+            0.002062438987195492,
+            -0.018551824614405632,
+            -0.0322159044444561,
+            0.016202466562390327,
+            -0.011112191714346409,
+            -0.018160264939069748,
+            0.015581372193992138,
+            -0.0046582091599702835,
+            -0.01875435560941696,
+            0.006875921972095966,
+            -0.01102442853152752,
+            -0.007999967783689499,
+            -0.032891009002923965,
+            -0.008675070479512215,
+            -0.0064809867180883884,
+            -0.004722343757748604,
+            -0.02198134735226631,
+            -0.01645900495350361,
+            -0.007115583401173353,
+            -0.03308003768324852,
+            -0.01184130273759365,
+            0.024303700774908066,
+            -0.025977954268455505,
+            -0.002192396204918623,
+            0.0008147646440193057,
+            -0.007493640761822462,
+            -0.004523188807070255,
+            -0.02377711981534958,
+            0.0015502047026529908,
+            0.005265801679342985,
+            0.0001731005759211257,
+            0.017080100253224373,
+            -0.002320665866136551,
+            0.004219392314553261,
+            0.005471708253026009,
+            0.0045670703984797,
+            -0.0063527170568704605,
+            -0.011503751389682293,
+            0.03189185634255409,
+            -0.008749332278966904,
+            -0.013947623781859875,
+            -0.02507331781089306,
+            -0.01740415021777153,
+            -0.0029670766089111567,
+            -0.025478379800915718,
+            0.03275598585605621,
+            -0.023358555510640144,
+            -0.02149527333676815,
+            -0.022669950500130653,
+            -0.0010708817280828953,
+            -0.0009054815745912492,
+            -0.06686218082904816,
+            -0.0032894383184611797,
+            0.0471491776406765,
+            0.002717288676649332,
+            -0.007243852596729994,
+            -0.0060759251937270164,
+            -0.17304234206676483,
+            0.003571293782442808,
+            0.006565374322235584,
+            -0.034133195877075195,
+            0.014528212137520313,
+            0.032404933124780655,
+            0.005920651368796825,
+            -0.0067746564745903015,
+            -0.011226959526538849,
+            0.004742597229778767,
+            0.014069141820073128,
+            -0.01625647582113743,
+            -0.02924545295536518,
+            0.002324041211977601,
+            -0.009620214812457561,
+            0.008648066781461239,
+            0.010794893838465214,
+            0.02198134735226631,
+            0.038345836102962494,
+            0.012104593217372894,
+            0.0300825797021389,
+            -0.023898638784885406,
+            -0.010558607056736946,
+            0.011233710683882236,
+            -0.005539218429476023,
+            -0.008033722639083862,
+            -0.0066362605430185795,
+            0.016040442511439323,
+            0.022183876484632492,
+            0.01177379209548235,
+            -0.016215968877077103,
+            0.014919771812856197,
+            0.019983042031526566,
+            -0.013103744946420193,
+            0.009404181502759457,
+            -0.008107984438538551,
+            0.0018109631491824985,
+            -0.038426849991083145,
+            -0.01892988197505474,
+            0.00391222070902586,
+            0.02830030769109726,
+            0.0030042072758078575,
+            0.028948407620191574,
+            -0.004168759565800428,
+            0.008378025144338608,
+            0.006437105126678944,
+            0.03256695717573166,
+            -0.027274152263998985,
+            0.023966148495674133,
+            0.00690967682749033,
+            0.04777027294039726,
+            -0.03046063706278801,
+            -0.00828351080417633,
+            -0.003682685550302267,
+            0.016296980902552605,
+            -0.004968756344169378,
+            0.005377193447202444,
+            0.014312178827822208,
+            0.014501207508146763,
+            -0.009606712497770786,
+            -0.016391495242714882,
+            -0.029002415016293526,
+            0.0016615967033430934,
+            -0.019091906026005745,
+            -0.017093602567911148,
+            -0.03572643920779228,
+            -0.014514709822833538,
+            0.036617573350667953,
+            -0.020334096625447273,
+            0.0013578004436567426,
+            0.00687929755076766,
+            -0.009626965969800949,
+            -0.0012379697291180491,
+            -0.00012278431677259505,
+            0.005515589844435453,
+            -0.0068556685000658035,
+            -0.016756050288677216,
+            0.028462333604693413,
+            0.04004709795117378,
+            -0.013576316647231579,
+            0.00774342892691493,
+            0.036644577980041504,
+            0.014474203810095787,
+            0.005599977448582649,
+            -0.014204162172973156,
+            -0.01344129629433155,
+            0.0007556931814178824,
+            0.007223599590361118,
+            -0.023736612871289253,
+            -0.024155177175998688,
+            0.014069141820073128,
+            -0.022170376032590866,
+            -0.018254779279232025,
+            -0.01442019548267126,
+            0.009431186132133007,
+            0.003510534530505538,
+            -0.008310515433549881,
+            -0.007689420599490404,
+            0.015081795863807201,
+            -0.0035915467888116837,
+            0.007561150938272476,
+            0.004459053743630648,
+            -0.02518133446574211,
+            0.006160312797874212,
+            0.01299572829157114,
+            0.018079252913594246,
+            0.007898702286183834,
+            0.03091970644891262,
+            0.039317984133958817,
+            0.001045565470121801,
+            -0.008837095461785793,
+            0.010673374868929386,
+            -0.004671711008995771,
+            0.026058968156576157,
+            -0.002253155456855893,
+            0.024708760902285576,
+            -0.004830360412597656,
+            -0.030811691656708717,
+            0.033215057104825974,
+            0.015162808820605278,
+            0.030649665743112564,
+            -0.013184756971895695,
+            0.01459572184830904,
+            0.010612615384161472,
+            0.00675777904689312,
+            0.007817690260708332,
+            -0.07085878401994705,
+            -0.023385560140013695,
+            0.020415108650922775,
+            0.01302273292094469,
+            -0.026234494522213936,
+            0.01225986611098051,
+            -0.006389847956597805,
+            0.034133195877075195,
+            -0.032026875764131546,
+            0.019631989300251007,
+            0.0029890176374465227,
+            -0.0021788941230624914,
+            -0.0052894302643835545,
+            -0.0027071621734648943,
+            0.004492809064686298,
+            -0.02057713270187378,
+            -0.01674254983663559,
+            0.001525732222944498,
+            -0.02227839082479477,
+            0.02956950105726719,
+            -0.003480154788121581,
+            -0.011004175059497356,
+            0.0019476715242490172,
+            0.00926240999251604,
+            0.006298708729445934,
+            0.017714696004986763,
+            -0.03510534390807152,
+            -0.007689420599490404,
+            0.02599145658314228,
+            -0.00995776616036892,
+            0.0034970324486494064,
+            -0.033242061734199524,
+            -0.002977203344926238,
+            0.0011502063134685159,
+            -0.0009839623235166073,
+            -0.003416019957512617,
+            -0.01564888283610344,
+            -0.03470028191804886,
+            0.040128108114004135,
+            0.0028708744794130325,
+            0.001292821834795177,
+            0.00225484324619174,
+            0.025856437161564827,
+            -0.01682356186211109,
+            -0.006727399304509163,
+            0.009451438672840595,
+            -0.038426849991083145,
+            0.016958581283688545,
+            -0.006150186527520418,
+            -0.0406951941549778,
+            -0.02550538256764412,
+            -0.009080132469534874,
+            -0.013205010443925858,
+            -0.006248076446354389,
+            0.0072371019050478935,
+            0.023709610104560852,
+            0.01892988197505474,
+            0.005097026005387306,
+            -0.0021653922740370035,
+            0.010511349886655807,
+            0.008053976111114025,
+            -0.0058497656136751175,
+            -0.021387256681919098,
+            0.029353469610214233,
+            0.03051464632153511,
+            0.01752566732466221,
+            -0.023183029145002365,
+            -0.037535715848207474,
+            -0.0032573710195720196,
+            -0.007628661580383778,
+            -0.010781391523778439,
+            0.01631048321723938,
+            -0.0300825797021389,
+            0.014555215835571289,
+            -0.052955061197280884,
+            -0.022224383428692818,
+            -0.0027696092147380114,
+            -0.016621030867099762,
+            -0.007750180084258318,
+            -0.004023612476885319,
+            0.0021417634561657906,
+            -0.00791220460087061,
+            -0.00926240999251604,
+            -0.018322288990020752,
+            0.022467421367764473,
+            0.026329008862376213,
+            -0.013036234304308891,
+            -0.029380472376942635,
+            0.012759442441165447,
+            -0.02063114009797573,
+            -0.0025113823357969522,
+            0.019240429624915123,
+            0.017107104882597923,
+            0.0049822586588561535,
+            0.006707146298140287,
+            -0.0006181409698911011,
+            0.001891975523903966,
+            -0.024047160521149635,
+            -0.0007232038769870996,
+            0.011760290712118149,
+            -0.028057271614670753,
+            -0.014811755158007145,
+            -0.07377523183822632,
+            0.007270856760442257,
+            -0.02227839082479477,
+            0.00450968649238348,
+            0.01979401335120201,
+            -0.03553741052746773,
+            0.015473355539143085,
+            -0.02561339922249317,
+            0.005498711951076984,
+            0.007540897931903601,
+            -0.03302602842450142,
+            -0.01829528622329235,
+            -0.017768705263733864,
+            -0.008141739293932915,
+            -0.005785631015896797,
+            0.0038683388847857714,
+            0.00904637761414051,
+            -0.007034570910036564,
+            0.011814299039542675,
+            -0.0029248828068375587,
+            -0.005343438591808081,
+            0.005893647205084562,
+            0.02198134735226631,
+            -0.0015780527610331774,
+            -0.013157753273844719,
+            -0.006190692540258169,
+            -0.02017207071185112,
+            0.01344129629433155,
+            -0.01841680333018303,
+            0.000581432250328362,
+            0.024708760902285576,
+            -0.0003734162310138345,
+            0.009539201855659485,
+            -0.004614327568560839,
+            -0.019172918051481247,
+            -0.026869090273976326,
+            0.03380914777517319,
+            0.010423586703836918,
+            0.0128674590960145,
+            0.03518635779619217,
+            -0.015297829173505306,
+            -0.042369451373815536,
+            0.02357458882033825,
+            -0.02624799683690071,
+            -0.030352620407938957,
+            -0.019496968016028404,
+            0.000544301641639322,
+            0.012469148263335228,
+            0.018740853294730186,
+            0.0036556816194206476,
+            0.001262442208826542,
+            0.012145099230110645,
+            -0.0107341343536973,
+            -0.020334096625447273,
+            -0.0001954633480636403,
+            -0.004111376125365496,
+            0.013657329604029655,
+            0.00848604179918766,
+            -0.0027291029691696167,
+            0.020671647042036057,
+            0.04326058551669121,
+            -0.008384776301681995,
+            -0.010592362843453884,
+            -0.01905140094459057,
+            0.010092786513268948,
+            -0.007918955758213997,
+            -0.0033366954885423183,
+            0.025059815496206284,
+            -0.021184725686907768,
+            -0.01548685785382986,
+            -0.006318962201476097,
+            -0.009458189830183983,
+            0.01483875885605812,
+            0.02873237431049347,
+            0.01590542122721672,
+            -0.026410020887851715,
+            0.0018818490207195282,
+            0.020496120676398277,
+            -0.012394887395203114,
+            0.026869090273976326,
+            0.011388983577489853,
+            0.009127389639616013,
+            0.009815994650125504,
+            -0.009505447000265121,
+            0.008364523760974407,
+            0.005468332674354315,
+            -0.02233240008354187,
+            -0.01887587457895279,
+            0.001643031369894743,
+            0.0021687676198780537,
+            -0.01859232969582081,
+            -0.0002548513002693653,
+            -0.011085188016295433,
+            0.008134988136589527,
+            0.0016514702001586556,
+            -0.004172135144472122,
+            -0.011409237049520016,
+            -0.03402518108487129,
+            0.009545953013002872,
+            0.008999120444059372,
+            0.006575501058250666,
+            -0.0203475970774889,
+            -0.0023341679479926825,
+            -0.024020157754421234,
+            -0.034079186618328094,
+            -0.0007839631289243698,
+            -0.018389800563454628,
+            0.0072776079177856445,
+            0.005917275790125132,
+            0.023210033774375916,
+            0.03505133464932442,
+            0.010774640366435051,
+            0.0086008096113801,
+            -0.009113887324929237,
+            -0.02484378218650818,
+            -0.005586475599557161,
+            -0.02045561373233795,
+            -0.02037460170686245,
+            -0.02279146946966648,
+            0.0182142723351717,
+            0.014487705193459988,
+            0.014055639505386353,
+            0.02835431694984436,
+            -0.008850597776472569,
+            0.033215057104825974,
+            0.040506165474653244,
+            0.0018295284826308489,
+            -0.022858979180455208,
+            0.013042985461652279,
+            -0.01689107157289982,
+            -0.020050551742315292,
+            -0.008540050126612186,
+            0.0017518916865810752,
+            0.011584763415157795,
+            -0.024074165150523186,
+            -0.024803275242447853,
+            -0.027706218883395195,
+            0.00696706073358655,
+            0.018713848665356636,
+            0.07269506901502609,
+            0.007223599590361118,
+            -0.028948407620191574,
+            0.014069141820073128,
+            0.011199954897165298,
+            0.018308786675333977,
+            0.004678462166339159,
+            0.017944231629371643,
+            -0.015135804191231728,
+            -0.03181084245443344,
+            -0.0018058998975902796,
+            -0.02268345281481743,
+            -0.02429019846022129,
+            -0.030811691656708717,
+            -0.0029282583855092525,
+            -0.023696107789874077,
+            -0.01253665890544653,
+            0.01585141383111477,
+            -0.01049109734594822,
+            -0.01115269772708416,
+            0.002455686451867223,
+            0.01795773394405842,
+            0.021657297387719154,
+            0.003377201734110713,
+            -0.02870536968111992,
+            0.008303764276206493,
+            0.05892296880483627,
+            0.0056134797632694244,
+            0.019186420366168022,
+            -0.032404933124780655,
+            0.008884352631866932,
+            -0.004604200832545757,
+            -0.03443024307489395,
+            -0.006187316961586475,
+            0.012388136237859726,
+            0.005404197610914707,
+            -0.009080132469534874,
+            -0.0026936600916087627,
+            0.005559471435844898,
+            0.009606712497770786,
+            0.002529947552829981,
+            0.01806575059890747,
+            -0.018119757995009422,
+            -0.053441137075424194,
+            -0.0017080099787563086,
+            -0.0009004183229990304,
+            -0.008756083436310291,
+            -0.02273746207356453,
+            -0.005441328510642052,
+        ],
+    }
 
-    Args:
-        dict_data (dict): data to be saved
-        index_name (str): index name
-    """
-    # Save the dictionary to OpenSearch
-    response = client.index(index=index_name, body=dict_data)
-
-    # Print the response
-    print(response)
-
-
-def create_index(index_name: str):
-    # Create an index with non-default settings.
-    # https://docs.aws.amazon.com/opensearch-service/latest/developerguide/sizing-domains.html#bp-sharding
-    # shards should correspond to 10-30GB where search latency is objective
-    # should be 30-50GB for write-heavy jobs like log analytics
-    # by default, each index == 5 primary shards + 1 replica/primary == 10x total shards
-    # replica shard == copy of a primary shard
-    # shards are distributed amongst nodes for resilience
-    # index > shards > multiple nodes (assuming you have more than 1)
-    # lower shard count == faster reads
-    # higher shard count == faster writes
-    index_body = {"settings": {"index": {"number_of_shards": 1}}}
-
-    try:
-        if client.indices.exists(index=index_name):
-            print(f"Index {index_name} already exists")
-            return
-        response = client.indices.create(index_name, body=index_body)
-        print("\nCreating index:")
-        print(response)
-    except Exception as e:
-        print(e)
+    os_client.save_to_index(job_parsed, index_name=index_name)
